requirements.txt
asgiref==3.6.0
Django==3.2.18
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.5.0
